Move xr-interop tests out of test_geobox
