+++ tests/test_geobox.py
@@ -2,7 +2,6 @@
 #
 # Copyright (c) 2015-2020 ODC Contributors
 # SPDX-License-Identifier: Apache-2.0
-from unittest.mock import MagicMock
 
 import numpy as np
 import pytest
@@ -10,7 +9,6 @@ from affine import Affine
 
 from odc import geo as geometry
 from odc.geo import CRS
-from odc.geo._xr_interop import _mk_crs_coord, xr_coords
 from odc.geo.geobox import (
     GeoBox,
     bounding_box_in_pixel_domain,
@@ -224,55 +222,6 @@ def test_geobox():
         bounding_box_in_pixel_domain(
             GeoBox(1, 1, mkA(0), epsg4326), GeoBox(2, 3, mkA(0), epsg3577)
         )
-
-
-def test_geobox_xr_coords():
-    A = mkA(0, scale=(10, -10), translation=(-48800, -2983006))
-
-    w, h = 512, 256
-    gbox = GeoBox(w, h, A, epsg3577)
-
-    cc = xr_coords(gbox, with_crs=False)
-    assert list(cc) == ["y", "x"]
-    assert cc["y"].shape == (gbox.shape[0],)
-    assert cc["x"].shape == (gbox.shape[1],)
-    assert "crs" in cc["y"].attrs
-    assert "crs" in cc["x"].attrs
-
-    cc = xr_coords(gbox, with_crs=True)
-    assert list(cc) == ["y", "x", "spatial_ref"]
-    assert cc["spatial_ref"].shape == ()
-    assert cc["spatial_ref"].attrs["spatial_ref"] == gbox.crs.wkt
-    assert isinstance(cc["spatial_ref"].attrs["grid_mapping_name"], str)
-
-    # with_crs should be default True
-    assert list(xr_coords(gbox)) == list(xr_coords(gbox, with_crs=True))
-
-    cc = xr_coords(gbox, with_crs="Albers")
-    assert list(cc) == ["y", "x", "Albers"]
-
-    # geographic CRS
-    A = mkA(0, scale=(0.1, -0.1), translation=(10, 30))
-    gbox = GeoBox(w, h, A, "epsg:4326")
-    cc = xr_coords(gbox, with_crs=True)
-    assert list(cc) == ["latitude", "longitude", "spatial_ref"]
-    assert cc["spatial_ref"].shape == ()
-    assert cc["spatial_ref"].attrs["spatial_ref"] == gbox.crs.wkt
-    assert isinstance(cc["spatial_ref"].attrs["grid_mapping_name"], str)
-
-    # missing CRS for GeoBox
-    gbox = GeoBox(w, h, A, None)
-    cc = xr_coords(gbox, with_crs=True)
-    assert list(cc) == ["y", "x"]
-
-    # check CRS without name
-    crs = MagicMock()
-    crs.projected = True
-    crs.wkt = epsg3577.wkt
-    crs.epsg = epsg3577.epsg
-    crs._crs = MagicMock()
-    crs._crs.to_cf.return_value = {}
-    assert _mk_crs_coord(crs).attrs["grid_mapping_name"] == "??"
 
 
 def test_gbox_boundary():
